Production-readiness remediation of preview API surface
Address the findings in PROD_READY_ISSUES.md across kernel, drivers,
services, builders, common, and the testing adapters, with matching
tests and documentation corrections.

Code:
- Kernel: UnregisterAsync disposes a removed driver on the disposal
  budget only, so caller cancellation cannot orphan it undisposed and
  uncounted; dispose helpers no longer take a cancellation token.
- Drivers: Api.StreamInterrupted added to IsTransientCode so mid-stream
  resets on /events and /stats stay retryable; graceful warn-and-fallback
  for a malformed DOCKER_MODEL_RUNNER_URL at pack init.
- Builders: WithVolume accepts Windows-container destinations (C:\app)
  via the shared drive-prefix-aware path check.
- Testing: compose session-label overlay sweep widened to 24h so it reaps
  only crash-orphans, never an in-use overlay.

Tests:
- Public-API production-readiness tests per area (Category=Unit) plus
  regression guards for the transient-code, kernel-cancellation, and
  Windows-container-path fixes.

Docs:
- Correct the net8.0 drop (single-target net10.0), FluentDockerNotSupportedException,
  64 MiB stdout cap, UnpauseAsync Unknown-then-throw contract, the xUnit
  skip-when-unavailable recipe, JsonHelper, and Docker API stream error codes.

Verified: make check green (5246/5246 unit tests, line 87.16% / branch
78.68% coverage, 0 warnings).

+++ FluentDocker/Model/Models/Inference/Completion.cs
@@ -41,6 +41,7 @@ namespace FluentDocker.Model.Models.Inference
     [JsonPropertyName("model")] public string? Model { get; set; }
 
     /// <summary>The prompt text.</summary>
+    /// <remarks>Only OpenAI's single string prompt form is modeled; string-array batch prompts are not.</remarks>
     [JsonPropertyName("prompt")] public string? Prompt { get; set; }
 
     /// <summary>The maximum number of tokens to generate.</summary>
@@ -59,7 +60,7 @@ namespace FluentDocker.Model.Models.Inference
     [JsonPropertyName("stop")] public IList<string>? Stop { get; set; }
 
     /// <summary>Sampling seed.</summary>
-    [JsonPropertyName("seed")] public int? Seed { get; set; }
+    [JsonPropertyName("seed")] public long? Seed { get; set; }
 
     /// <summary>
     /// Pass-through for any OpenAI-compatible request field not modeled above (e.g.
